==> r/09_control-plots-2023/02_FSSgam_Length.R <==
###
# Project: mac - South-west Corner
# Data:    BRUV fish and habitat, broad bathymetry derivatives
# Task:    Run new models using updated legal sizes and size of maturity
# author:  Claude
# date:    March 2023
##

library(rstanarm)
library(tidyverse)
library(dplyr)
library(mgcv)
library(FSSgam)
library(MuMIn)
library(doBy)
library(GlobalArchive)
library(googlesheets4)

rm(list=ls())

# Set the study name
name <- '2020-2021_south-west_BOSS-BRUV_length' # for the study

dat <- readRDS('data/tidy/2023-03_dat.length-maturity.rds')%>%
  # dplyr::filter(scientific%in%c("greater than legal size","smaller than legal size"))%>%
  glimpse()

unique(dat$scientific)

test <- dat %>%
  group_by(sample)%>%
  dplyr::summarise(n=n())

unique(dat$scientific)
# Set predictor variables 
pred.vars=c("mean.relief","detrended","broad.macroalgae","broad.reef",
            "distance.to.ramp", "tpi","roughness","depth")

unique.vars=unique(as.character(dat$scientific))
unique.vars

unique.vars.use=character()
for(i in 1:length(unique.vars)){
  temp.dat=dat[which(dat$scientific==unique.vars[i]),]
  if(length(which(temp.dat$number==0))/nrow(temp.dat)<0.8){
    unique.vars.use=c(unique.vars.use,unique.vars[i])}
}

unique.vars.use  

# Run the full subset model selection----
savedir <- "output/2023_fish-gamms"
resp.vars=unique.vars.use
use.dat=as.data.frame(dat)
str(use.dat)

factor.vars=c("status")# Status as a Factor with two levels
out.all=list()
var.imp=list()

# Loop through the FSS function for each Taxa----
for(i in 1:length(resp.vars)){
  use.dat=as.data.frame(dat[which(dat$scientific==resp.vars[i]),])
  use.dat$site <- as.factor(use.dat$site)
  Model1=gam(number~s(depth,k=3,bs='cr')+s(site,bs='re'),                                       
             family=tw(),  data=use.dat)
  
  model.set=generate.model.set(use.dat=use.dat,
                               test.fit=Model1,
                               factor.smooth.interactions = FALSE,
                               # smooth.smooth.interactions = c("depth"),
                               pred.vars.cont=pred.vars,
                               pred.vars.fact=factor.vars,
                               #cyclic.vars = cyclic.vars,
                               #linear.vars="depth",
                               k=3,
                               null.terms="s(site ,bs='re')"                           
  )
  out.list=fit.model.set(model.set,
                         max.models=600,
                         parallel=T,n.cores = 8)                                #change for computer here
  names(out.list)
  
  out.list$failed.models # examine the list of failed models
  mod.table=out.list$mod.data.out  # look at the model selection table
  mod.table=mod.table[order(mod.table$AICc),]
  mod.table$cumsum.wi=cumsum(mod.table$wi.AICc)
  out.i=mod.table[which(mod.table$delta.AICc<=2),]
  out.all=c(out.all,list(out.i))
  # var.imp=c(var.imp,list(out.list$variable.importance$aic$variable.weights.raw)) #Either raw importance score
  var.imp=c(var.imp,list(out.list$variable.importance$aic$variable.weights.raw)) #Or importance score weighted by r2
  
  # plot the best models
  for(m in 1:nrow(out.i)){
    best.model.name=as.character(out.i$modname[m])
    
    png(file = paste(savedir, paste(name, m, resp.vars[i], "mod_fits.png", sep = "_"), sep = "/"))
    if(best.model.name!="null"){
      par(mfrow=c(3,1),mar=c(9,4,3,1))
      best.model=out.list$success.models[[best.model.name]]
      plot(best.model,all.terms=T,pages=1,residuals=T,pch=16)
      mtext(side=2,text=resp.vars[i],outer=F)}  
    dev.off()
  }
}

# Model fits and importance---
names(out.all)=resp.vars
names(var.imp)=resp.vars
all.mod.fits=do.call("rbind",out.all)
all.var.imp=do.call("rbind",var.imp)
write.csv(all.mod.fits[ , -2], file = paste(savedir, paste(name, "all.mod.fits.csv", sep = "_"), sep = "/"))
write.csv(all.var.imp, file = paste(savedir, paste(name, "all.var.imp.csv", sep = "_"), sep = "/"))
